Moved LoginTracker class into AuthClient class
