+++ jenjin-core-client/src/test/java/com/jenjinstudios/client/message/ExecutableLogoutResponseTest.java
@@ -1,7 +1,6 @@
 package com.jenjinstudios.client.message;
 
 import com.jenjinstudios.client.net.AuthClient;
-import com.jenjinstudios.client.net.LoginTracker;
 import com.jenjinstudios.core.io.Message;
 import com.jenjinstudios.core.io.MessageRegistry;
 import org.testng.annotations.Test;
@@ -10,6 +9,7 @@ import static org.mockito.Mockito.*;
 
 /**
  * Test the ExecutableLogoutResponse class.
+ *
  * @author Caleb Brinkman
  */
 public class ExecutableLogoutResponseTest
@@ -18,18 +18,18 @@ public class ExecutableLogoutResponseTest
      * Test the execution of the ExecutableLogoutResponse.
      */
     @Test
-	public void testMessageExecution() {
-		Message loginResponse = MessageRegistry.getInstance().createMessage("LogoutResponse");
-		loginResponse.setArgument("success", true);
+    public void testMessageExecution() {
+        Message loginResponse = MessageRegistry.getInstance().createMessage("LogoutResponse");
+        loginResponse.setArgument("success", true);
 
-		AuthClient authClient = mock(AuthClient.class);
-		LoginTracker loginTracker = mock(LoginTracker.class);
-		when(authClient.getLoginTracker()).thenReturn(loginTracker);
+        AuthClient authClient = mock(AuthClient.class);
+        AuthClient.LoginTracker loginTracker = mock(AuthClient.LoginTracker.class);
+        when(authClient.getLoginTracker()).thenReturn(loginTracker);
 
         ExecutableLogoutResponse response = new ExecutableLogoutResponse(authClient, loginResponse);
         response.runImmediate();
         response.runDelayed();
 
-		verify(loginTracker).setLoggedIn(false);
-	}
+        verify(loginTracker).setLoggedIn(false);
+    }
 }
